シーンマネージャの実装
