+++ DirectXGame/Scene/GameBaseScene.cpp
@@ -1,0 +1,1 @@
+#include "GameBaseScene.h"
