+++ DirectXGame/main.cpp
@@ -1,11 +1,10 @@
 #define DIRECTINPUT_VERSION    0x0800 //DirectInputのバージョン指定
-#include "GameScene.h"
-#include "SIFrameWork.h"
+#include "Game.h"
 
 int WINAPI WinMain(HINSTANCE, HINSTANCE, LPSTR, int)
 {
-	SIFrameWork* game = new GameScene();
-	
+	SIFrameWork* game = new Game();
+
 	game->Run();
 
 	delete game;
